% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTitles.R
\name{addTitles}
\alias{addTitles}
\title{get the titles of a bunch of GSM entries}
\usage{
addTitles(x, column = "subject", cachePath = NULL)
}
\arguments{
\item{x}{either a grSet from sesamizeGEO, or a bunch of GSMs}

\item{column}{name of the column holding the GSMs ("subject")}

\item{cachePath}{where to cache the GEOmetadb sqlite file (tempdir())}
}
\value{
\preformatted{       a retitled grSet (if x is a grSet) or a bunch of titles
}
}
\description{
like it says on the tin. this is usually for labeling purposes.
}
