% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asMethBlocks.R
\name{asMethBlocks}
\alias{asMethBlocks}
\title{Collapse a GenomicRatioSet or similar over methylation blocks (per Kaplan)}
\usage{
asMethBlocks(x, genome = "hg19", proper = FALSE, stable = FALSE)
}
\arguments{
\item{x}{a GenomicRatioSet or something like it}

\item{genome}{genome to use for methylation block coordinates (hg19)}

\item{proper}{restrict to methylation blocks defined in Kaplan? (FALSE)}

\item{stable}{restrict to methylation blocks shared by hg19&hg38? (FALSE)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     an object with same colData but with new rowRanges & assays
}\if{html}{\out{</div>}}
}
\description{
Collapse a GenomicRatioSet or similar over methylation blocks (per Kaplan)
}
\details{
By default, data(methBlocks) is used for summarization, singleton probes are
included, methylation block rownames are returned against hg19, and neither
singletons (!proper) nor hg19-only blocks (!stable) are filtered out.
}
