% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assayToBigWigs.R
\name{assayToBigWigs}
\alias{assayToBigWigs}
\title{Dump data as bigWig tracks for (e.g.) igv.js (which can't handle multiGCTs)}
\usage{
assayToBigWigs(x, assay = 1, genome = NULL, onlycg = TRUE)
}
\arguments{
\item{x}{a RangedSummarizedExperiment-like object of some sort}

\item{assay}{name or index of an assay to dump (default: assays(x)[\link{1}])}

\item{genome}{(optional) genome to use (default: look in genome(x))}

\item{onlycg}{(optional) only dump CpG probes (for methylation)? (TRUE)}
}
\value{
list     a list of filenames dumped
}
\description{
Dump data as bigWig tracks for (e.g.) igv.js (which can't handle multiGCTs)
}
