% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseAt.R
\name{collapseAt}
\alias{collapseAt}
\title{Collapse a SummarizedExperiment-like object over a GRanges}
\usage{
collapseAt(x, y, how = c("medians", "means"), minprobes = 2, imp = TRUE)
}
\arguments{
\item{x}{anything descended from a SummarizedExperiment}

\item{y}{regions to summarize over, as a GRanges}

\item{how}{medians or means? (medians)}

\item{minprobes}{minimum probes to summarize over (1)}

\item{imp}{impute any remaining NAs? (FALSE)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     an object with same colData but with new rowRanges & assays
}\if{html}{\out{</div>}}
}
\description{
Collapse a SummarizedExperiment-like object over a GRanges
}
\details{
No imputation is performed and only unique ranges are summarized.
If you want to disjoin ranges, do it before running this function.
The function just takes whatever is in assays(x)[\link{1}] and summarizes that.
}
