% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_metrics.R
\name{control_metrics}
\alias{control_metrics}
\title{bead array control processing function: get controls from red/green mats}
\usage{
control_metrics(rgSet, dft = NULL, baseline = 3000, biotin.baseline = 1)
}
\arguments{
\item{rgSet}{an rgSet with control probe intensities and annotation}

\item{dft}{optional single-row data.frame of thresholds (default)}

\item{baseline}{offset for Extension green based background (3000)}

\item{biotin.baseline}{baseline offset for biotin staining probes (1)}
}
\value{
\preformatted{           a matrix (cols = metrics, rows = samples)
}
}
\description{
mostly lifted from Sean Maden's recountmethylation supplement
https://github.com/metamaden/recountmethylationManuscriptSupplement
modified slightly to take advantage of rgSet annotations and to Go Big,
i.e. this works reasonably well on fairly large rgSets (like all of GEO)
}
