% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpgAnnoByChr.R
\name{cpgAnnoByChr}
\alias{cpgAnnoByChr}
\title{DMRcate doesn't play nicely with HDF5-backed SEs, so I wrote this instead.}
\usage{
cpgAnnoByChr(object, design, coef = 2, fdr = 0.05, ...)
}
\arguments{
\item{object}{a GenomicRatioSet}

\item{design}{a design matrix}

\item{coef}{which column to fit (default is 2)}

\item{fdr}{FDR cutoff (0.05)}

\item{...}{other arguments, to be passed (in principle) to DMRcate}
}
\value{
\preformatted{   annotated fits for x and its design matrix
}
}
\description{
TODO: run in parallel, especially for HDF5-backed objects.
}
