% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpFromIDATs.R
\name{dumpFromIDATs}
\alias{dumpFromIDATs}
\title{read in a directory's worth of IDATs, flag controls, sesamize, then dump}
\usage{
dumpFromIDATs(stub, path = ".", ...)
}
\arguments{
\item{stub}{mandatory prefix for files (stub_betas.csv, stub_qc.csv)}

\item{path}{where to save the dumped CSV files (".")}

\item{...}{options to pass to sesame::sesamize}
}
\value{
a GenomicRatioSet (or an rgSet in case of failure)
}
\description{
for ease of factorization, inferred sex and flagged controls are written out
as a control/QC metrics matrix, along with the fraction of NAs among probes
that would not typically be masked by sesame, and the median total intensity
}
