% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elts.R
\name{elts}
\alias{elts}
\title{get elements from a consistently separated character string}
\usage{
elts(x, y = "_", z = 1, naify = FALSE)
}
\arguments{
\item{x}{the string}

\item{y}{the separator}

\item{z}{the element\link{s} to get}

\item{naify}{return NA_character_ if y is not present? (FALSE)}
}
\value{
\preformatted{   a string, joined by the separator if z is a vector
}
}
\description{
Another function that I keep rewriting; no more!
}
\examples{

string <- "foo_123_bar.baz.qux"

elts(string)               # [1] "foo" 
elts(string, z=2:3)        # [1] "123_bar.baz.qux"
elts(string, y=".", z=2:3) # [1] "baz.qux"

strings <- paste0(letters[1:3], "_string")

elts(strings)              # a_string b_string c_string 
                           #      "a"      "b"      "c" 

strings2 <- c(strings, "nothing") 

elts(strings2)             #  a_string  b_string  c_string   nothing 
                           #       "a"       "b"       "c" "nothing" 
elts(strings2, naify=TRUE) #  a_string b_string c_string  nothing 
                           #       "a"      "b"      "c"       NA 

}
