% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftpGSMs.R
\name{ftpGSMs}
\alias{ftpGSMs}
\title{A slower way to get the IDAT files if for some reason
the GEOmetadb.sqlite file is out of date or stale.}
\usage{
ftpGSMs(GSMs = NULL, cachePath = NULL)
}
\arguments{
\item{GSMs}{a vector of GEO sample IDs}

\item{cachePath}{where the cached GEOmetadb SQLite file lives (tempdir())}
}
\value{
FTP file paths for IDATs
}
\description{
A slower way to get the IDAT files if for some reason
the GEOmetadb.sqlite file is out of date or stale.
}
\examples{
ftpGSMs("GSM3911862")
}
