% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBarcodes.R
\name{getBarcodes}
\alias{getBarcodes}
\title{retrieve all the IDAT Basenames in a directory}
\usage{
getBarcodes(dir = ".", elt = 1:2, ...)
}
\arguments{
\item{dir}{the path to the directory (default: use ".", the current path)}

\item{elt}{which fragment\link{s} to use as the barcode (default: 1 and 2)}

\item{...}{other arguments to pass to elts()}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ the barcodes
}\if{html}{\out{</div>}}
}
\description{
Note that getBarcodes(dir, 1) will get the chip address for unmolested IDATs.
}
