% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGSMs.R
\name{getGSMs}
\alias{getGSMs}
\title{fetch supplemental data (IDATs) for a given dataset}
\usage{
getGSMs(GSMs, cachePath = NULL)
}
\arguments{
\item{GSMs}{a vector of GEO sample IDs}

\item{cachePath}{where the cached GEOmetadb SQLite file lives (tempdir())}
}
\value{
\preformatted{       a list of supplementary filenames, usually 2 per GSM
}
}
\description{
There's probably a better way to do this, but we haven't found it.
IDAT files will be dumped into the working directory, so setwd() first.
}
