% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDAT.R
\name{getIDAT}
\alias{getIDAT}
\title{download IDAT file from a provided URL}
\usage{
getIDAT(IDAT)
}
\arguments{
\item{IDAT}{URL for IDAT file to attempt downloading}
}
\value{
\preformatted{basename of IDAT file (see R basename function for more info)
}
}
\description{
Checks to see if the requested IDAT file exists in the current working directory.
Downloads the IDAT if it does not exist. Returns basename of IDAT file.
}
