% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhenoData.R
\name{getPhenoData}
\alias{getPhenoData}
\title{Retrieve phenoData data frame from GEO series}
\usage{
getPhenoData(series, gse, gpl)
}
\arguments{
\item{series}{GEO series accessed by GEOquery::getGEO()}

\item{gse}{String with the GEO series name}

\item{gpl}{String with the GEO platform name}
}
\value{
phenoData data frame
}
\description{
Given a GEO series accessed by GEOquery::getGEO(), return the phenoData data
frame
}
\examples{

}
