% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRGChannelSet.R
\name{getRGChannelSet}
\alias{getRGChannelSet}
\title{Assemble a suitable RGChannelSet for sesamizing}
\usage{
getRGChannelSet(subjects = NULL, frags = 1:3, samps = NULL,
  rightType = NULL, force = FALSE)
}
\arguments{
\item{subjects}{The names of each subject, or NULL to autodetect
(DEFAULT: NULL)}

\item{frags}{Which elements to extract for the array Basename (DEFAULT: 1:3)}

\item{samps}{Targets data.frame - overrides subjects and frags
(DEFAULT: NULL)}

\item{rightType}{"Correct" array type, tries to guess is NULL (DEFAULT: NULL)}

\item{force}{Parse IDAT files with different array size,
see the man page ?read.metharray (DEFAULT: FALSE)}
}
\value{
An RGChannelSet with pData $subject and $Basename filled
}
\description{
Note: this should probably go away in favor of signalSet parallel processing
}
