% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSamps.R
\name{getSamps}
\alias{getSamps}
\title{Helper function to read in subjects (samples) and Basename columns}
\usage{
getSamps(subjects = NULL, frags = 1:3, check.frags = TRUE)
}
\arguments{
\item{subjects}{The names of each subject, will autodetect if NULL
(DEFAULT: NULL)}

\item{frags}{Which elements to extract for the array Basename, ignored if
\code{check.frags} = TRUE (DEFAULT: 1:3)}

\item{check.frags}{Check if frags covers all elements up to "idat.gz"
(DEFAULT: TRUE)}
}
\value{
A data.frame with Basename and subject columns
}
\description{
Helper function to read in subjects (samples) and Basename columns
}
