% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelAnnotations.R
\name{labelAnnotations}
\alias{labelAnnotations}
\title{label ComplexHeatmap annotations}
\usage{
labelAnnotations(annoTop = NULL, annoBot = NULL, fontsize = 11)
}
\arguments{
\item{annoTop}{top annotations}

\item{annoBot}{bottom annotations}

\item{fontsize}{font size (default 11)}
}
\description{
label ComplexHeatmap annotations
}
