% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAsHDF5.R
\name{loadAsHDF5}
\alias{loadAsHDF5}
\title{EXPERIMENTAL harness to eventually bypass in-core processing for HDF5}
\usage{
loadAsHDF5(dir)
}
\arguments{
\item{dir}{name of the directory where the object and its data are stored}
}
\value{
\preformatted{ HDF5-backed GenomicRatioSet (or SummarizedExperiment, or w/e)
}
}
\description{
Right now, all this does is load the GenomicRatioSet in an HDF5 directory.
}
