% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_metadata.R
\name{merge_metadata}
\alias{merge_metadata}
\title{convenience function for merging large SummarizedExperiment-like objects}
\usage{
merge_metadata(x, y)
}
\arguments{
\item{x}{the first object}

\item{y}{the second object}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a list() with the merged metadata
}\if{html}{\out{</div>}}
}
\description{
The two objects' metadata must have identical names,
and each metadata items' rownames must be identical.
}
