% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rgSets.R
\name{merge_rgSets}
\alias{merge_rgSets}
\title{convenience function for merging large RGChannelSets}
\usage{
merge_rgSets(rgSet1, rgSet2)
}
\arguments{
\item{rgSet1}{the first RGChannelSet}

\item{rgSet2}{the second RGChannelSet}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   the result of cbind(rgSet1, rgSet2) with sensible metadata 
}\if{html}{\out{</div>}}
}
\description{
Merges metadata, drops from each RGChannelSet, merges RGChannelSets, re-adds
RGChannelSets' metadata must have identical names, and platform must match.
}
