% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMeth.R
\name{mergeMeth}
\alias{mergeMeth}
\title{helper function to merge grSets}
\usage{
mergeMeth(x, y)
}
\arguments{
\item{x}{a GenomicRatioSet}

\item{y}{another GenomicRatioSet}
}
\value{
a GenomicRatioSet combining the two
}
\description{
helper function to merge grSets
}
