% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeWithSNPs.R
\name{mergeWithSNPs}
\alias{mergeWithSNPs}
\title{helper function to merge grSets while keeping their SNP matrices intact}
\usage{
mergeWithSNPs(x, y)
}
\arguments{
\item{x}{a GenomicRatioSet}

\item{y}{another GenomicRatioSet}
}
\value{
a GenomicRatioSet with sensibly managed SNPs in its metadata()
}
\description{
helper function to merge grSets while keeping their SNP matrices intact
}
