% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFlaggedControls.R
\name{plotFlaggedControls}
\alias{plotFlaggedControls}
\title{simple function for plotting control probe failure flaggings}
\usage{
plotFlaggedControls(x, ...)
}
\arguments{
\item{x}{an object where !is.null(metadata(x)$control_flagged), or a matrix}

\item{...}{additional parameters to pass to ComplexHeatmap::Heatmap}
}
\value{
a Heatmap
}
\description{
If an object is provided without suitable metadata, this function will fail.
See flag_control_failures().
}
\seealso{
flag_control_failures
}
