% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSNPcalls.R
\name{plotSNPcalls}
\alias{plotSNPcalls}
\title{call and plot SNPs from a matrix of SNP betas}
\usage{
plotSNPcalls(x, rotate = FALSE, ...)
}
\arguments{
\item{x}{a grSet with SNPs in its metadata(), or a SNP matrix}

\item{rotate}{rotate the subjects onto the side? (FALSE)}

\item{...}{other arguments passed on to Heatmap}
}
\description{
plot SNPs living in metadata(x)$SNPs, provided they match sampleNames(x)
}
