% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSNPs.R
\name{plotSNPs}
\alias{plotSNPs}
\title{another oldie but goodie}
\usage{
plotSNPs(x, rotate = FALSE, ...)
}
\arguments{
\item{x}{a grSet with SNPs in its metadata(), or a SNP matrix}

\item{rotate}{rotate the subjects onto the side? (FALSE)}

\item{...}{other arguments passed on to Heatmap}
}
\description{
plot SNPs living in metadata(x)$SNPs, provided they match sampleNames(x)
}
