% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotXYstats.R
\name{plotXYstats}
\alias{plotXYstats}
\title{simple function for plotting XY stats (split by inferred sex if found)}
\usage{
plotXYstats(grSet, column = NULL, ...)
}
\arguments{
\item{grSet}{a GenomicRatioSet}

\item{column}{optional column name with sex information}

\item{...}{additional parameters to pass to ComplexHeatmap::Heatmap}
}
\value{
a Heatmap
}
\description{
simple function for plotting XY stats (split by inferred sex if found)
}
