% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dmrcate.output.R
\name{print.dmrcate.output}
\alias{print.dmrcate.output}
\title{since I use DMRcate a lot, I wrote this to avoid borking the command line}
\usage{
\method{print}{dmrcate.output}(object)
}
\arguments{
\item{object}{an object of class dmrcate.output}
}
\value{
\preformatted{   nothing, just cat the object's various properties
}
}
\description{
since I use DMRcate a lot, I wrote this to avoid borking the command line
}
