% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFromIDATs.R
\name{processFromIDATs}
\alias{processFromIDATs}
\title{read in a directory's worth of IDATs, flag controls, and sesamize them all}
\usage{
processFromIDATs(..., frags = 1:3, addgeo = FALSE)
}
\arguments{
\item{...}{options to pass to sesame::sesamize}

\item{frags}{which elements of the filename are relevant? (1:3)}

\item{addgeo}{optional: try to annotate from GEO? (FALSE)}
}
\value{
a GenomicRatioSet (or an rgSet in case of failure)
}
\description{
Note: this function simply reads all IDATs in the current directory,
runs QC, stuffs that into metadata, and sesamizes the result.
}
