% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processRgSet.R
\name{processResult}
\alias{processResult}
\title{sesamize a result with sensible QC and passing BPPARAM, etc.}
\usage{
processResult(x, addgeo = FALSE, justResult = TRUE, ...)
}
\arguments{
\item{x}{a SummarizedExperiment-derived object with SNPs and controls}

\item{addgeo}{optional: try to annotate from GEO, if not already done? (F)}

\item{...}{options to pass to sesamize}

\item{justDump}{just dump the sesamized result? (TRUE)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a GenomicRatioSet (or an res if failure)
}\if{html}{\out{</div>}}
}
\description{
sesamize a result with sensible QC and passing BPPARAM, etc.
}
