% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processRgSet.R
\name{processRgSet}
\alias{processRgSet}
\title{sesamize an RGChannelSet with sensible QC and passing BPPARAM, etc.}
\usage{
processRgSet(rgSet, addgeo = FALSE, ...)
}
\arguments{
\item{rgSet}{an RGChannelSet with SNPs and control_flagged in metadata()}

\item{addgeo}{optional: try to annotate from GEO, if not already done? (F)}

\item{...}{options to pass to sesamize}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a GenomicRatioSet (or an rgSet if failure)
}\if{html}{\out{</div>}}
}
\description{
sesamize an RGChannelSet with sensible QC and passing BPPARAM, etc.
}
