% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purify.R
\name{purify}
\alias{purify}
\alias{purifyBetas}
\title{The world's dumbest deconvolution function}
\usage{
purify(x, purity, sqz = 1e-06, betas = TRUE)
}
\arguments{
\item{x}{a SummarizedExperiment-like object, or a matrix of values}

\item{purity}{a vector of estimated purities (\code{sqz} to 1.00000, inclusive)}

\item{sqz}{a squeeze factor to avoid +/-Inf values in \code{flogit} (1e-6)}

\item{betas}{are the values of assay(x) between 0 and 1? (default is TRUE)}
}
\value{
\preformatted{   whatever `x` is (matrix or SE), purified (see Details)
}
}
\description{
Given x with rows and columns, and purity estimates for each column,
\code{purifyBetas} takes (squeezed) logits of assay(x), divides by \code{purity}
to push values towards the extremes, and then takes (de-squeezed)
expits to put the purified values back on a 0-1 scale.
}
\details{
\code{purify} is even dumber and simply divides each assays(x) by purity,
unless the argument \code{betas} is TRUE (the default), in which case it
calls \code{purifyBetas} instead. (This is a methylation-centric package,
although VAFs can be treated the same way, hence the default setting.)\preformatted{  If you pass `purity` estimates of 0, you'll get an error.
           If `x` is a GenomicRatioSet, only assays(x)$Beta is purified.
           No attempt is made to preserve imprinting or any such thing,
           although the tendency of imprinted regions to fall near zero
           on a logit scale, and the fact that zero times anything is 
           still zero, does incidentally encourage this. However, any
           resemblance to biological sanity is purely coincidental.
}
}
