% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purifyBetas.R
\name{purifyBetas}
\alias{purifyBetas}
\title{The world's dumbest deconvolution function}
\usage{
purifyBetas(x, purity, sqz = 1e-06)
}
\arguments{
\item{x}{a GenomicRatioSet (if passed a matrix, will treat as betas)}

\item{purity}{a vector of estimated purities (\code{sqz} to 1.00000, inclusive)}

\item{sqz}{a squeeze factor to avoid +/-Inf values in \code{flogit} (1e-6)}
}
\value{
\preformatted{   whatever `x` is (matrix or grSet), with purified beta values 
}
}
\description{
Given estimated purities for each sample, and a GenomicRatioSet,
\code{purify} takes (squeezed) logits of beta values, divides by purity
to push values towards the extremes, and then takes (de-squeezed)
expits to put the purified values back on a 0-1 scale. No attempt
is made to preserve imprinting or any such thing, although the
tendency of imprinted regions to end up near zero on a logit scale
and the tendency of zeroes multiplied by anything to be zeroes does
incidentally encourage this. However, any resemblance to biological
sanity is purely coincidental. Also, we don't care how you estimate
purity. Spin the bottle, use flow cytometry, whatever you like. All
this function does is take logits, divide, and take expits.
}
