% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMethArrays.R
\name{readMethArrays}
\alias{readMethArrays}
\title{Read Illumina methylation arrays}
\usage{
readMethArrays(base, targets, extended = FALSE, verbose = FALSE,
  force = FALSE, dropWrongType = TRUE, rightType = NULL)
}
\arguments{
\item{base}{Base directory of IDAT files}

\item{targets}{data.frame of target Red/Green IDAT files}

\item{extended}{Return a RGChannelSetExtended}

\item{verbose}{Output extra information}

\item{force}{Parse IDAT files with different array size, see the man
page ?read.metharray for more information}

\item{dropWrongType}{Drop "wrong" array types (DEFAULT: TRUE)}

\item{rightType}{"Correct" array type, tries to guess if NULL
(DEFAULT: NULL)}
}
\value{
A RGChannelSet
}
\description{
Code based on minfi::read.metharray and minfi::read.metharray.exp. Additional
functionality was needed beyond the minfi function as it is possible to find
datasets which have been incorrectly labeled as 450k or EPIC arrays.
}
\examples{

}
\seealso{
minfi::read.metharray() and minfi::read.metharray.exp()
}
