% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressFactors.R
\name{regressFactors}
\alias{regressFactors}
\title{run (e.g.) eBayes(lmFit(assay(x), design)) on columns of (e.g.) assay(x)}
\usage{
regressFactors(
  x,
  columns,
  y = NULL,
  how = c("lmFit", "concordance"),
  check = TRUE,
  purify = FALSE,
  purity = NULL,
  betas = FALSE,
  ...
)
}
\arguments{
\item{x}{a SummarizedExperiment-like object}

\item{columns}{colData(x) column names to regress as predictors}

\item{y}{optional matrix w/ ncols(x) columns (default is assay(x))}

\item{how}{"lmFit" (default) or "concordance" (survival::concordance)}

\item{check}{ensure that the selected columns of y are non-NA? (TRUE)}

\item{purify}{purify \code{x}? (default is FALSE, but this can be useful)}

\item{purity}{if \code{purify} == TRUE, a required vector of purities for x}

\item{betas}{if \code{purify} == TRUE, are the factors beta values? (FALSE)}

\item{...}{additional arguments to lmFit, if how == "fit"}
}
\value{
\preformatted{       a fitted model (use topTable on it) or concordance object
}
}
\description{
if \code{y} is specified, use that instead of assay(x) but constrain it the same
if \code{purify} is specified, purify assay(x) before proceeding with the fit(s)
}
\seealso{
\preformatted{      purify
}

\preformatted{      limma::lmFit
}

\preformatted{      limma::topTable
}

\preformatted{      survival::concordance
}
}
