% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_meta.R
\name{rename_meta}
\alias{rename_meta}
\title{convenience function to rename columns of *Set metadata matrices}
\usage{
rename_meta(x)
}
\arguments{
\item{x}{the SummarizedExperiment-derived thing with metadata to rename}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ the same, but with metadata columns renamed to colnames(x)
}\if{html}{\out{</div>}}
}
\description{
convenience function to rename columns of *Set metadata matrices
}
