% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowDataToBigWig.R
\name{rowDataToBigWig}
\alias{rowDataToBigWig}
\title{convenience function to dump (e.g.) W matrix values to bigWig files}
\usage{
rowDataToBigWig(x, mcol, fname = NULL, seqs = NULL, nmf = NULL)
}
\arguments{
\item{x}{a GRanges, a SummarizedExperiment-like object, or its rowRanges}

\item{mcol}{the name of the mcols column in x or rowData(x) to bigWigify}

\item{fname}{the name of the bigWig file (else default to mcol.genome.bw)}

\item{seqs}{optional seqinfo for any \code{x} without seqlengths}

\item{nmf}{NMF object with rownames identical to x, for its W matrix}
}
\description{
convenience function to dump (e.g.) W matrix values to bigWig files
}
