% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowExtremalities.R
\name{rowExtremalities}
\alias{rowExtremalities}
\title{retrieve row-wise ratios of the actual SDs vs theoretical max (Bernoulli SD)}
\usage{
rowExtremalities(x)
}
\arguments{
\item{x}{a Matrix, DelayedArray, or GenomicRatioSet}
}
\value{
\preformatted{ actualSd/bernoulliSd for each row
}
}
\description{
This is useful in finding probes/features that discriminate classes well.
}
