% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesamiser.R
\name{sesamiser}
\alias{sesamiser}
\title{like getGEO, if getGEO worked directly on IDATs and masked off deletions}
\usage{
sesamiser(
  GSE,
  element = 1,
  suppcols = c("supplementary_file", "supplementary_file.1"),
  path = "."
)
}
\arguments{
\item{GSE}{which GSE to get}

\item{element}{which element of the getGEO result to work on? (1)}

\item{suppcols}{cols ("supplementary_file", "supplementary_file.1")}
}
\value{
\preformatted{             a sesamized GenomicRatioSet
}
}
\description{
Note 0: this function will download the data with getGEO first (!)
Note 1: you may fare better with recountmethylation, check there first.
Note 2: this would work just as well off the moby recountmethylation rgset.
Note 3: this function will exit with an error if no IDATs are available.
Note 4: if there is one supplementary file column, we will need to split it.
Note 5: there are some mysterious new errors from sesame::sesamize() lately.
}
