% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesamize.R
\name{sesamize}
\alias{sesamize}
\title{"fix" an RGset (for which IDATs may be unavailable) with Sesame}
\usage{
sesamize(x, naFrac = 1, parallel = FALSE, mft = NULL)
}
\arguments{
\item{x}{an RGChannelSet, perhaps with colData of various flavors}

\item{naFrac}{maximum NA fraction for a probe before it gets dropped (1)}

\item{parallel}{attempt to run in parallel? (This is a bad idea on laptops)}

\item{mft}{a data.frame with columns Probe_ID, M, U, and col (guess)}
}
\value{
a sesamized GenomicRatioSet from the input RGChannelSet
}
\description{
"fix" an RGset (for which IDATs may be unavailable) with Sesame
}
\examples{
\donttest{

  # reprocess some 450k data from an existing RGChannelSet 
  if (require(FlowSorted.CordBloodNorway.450k) && 
    require(IlluminaHumanMethylation450kmanifest) && 
    require(IlluminaHumanMethylation450kanno.ilmn12.hg19)) {
    sesamized <- sesamize(FlowSorted.CordBloodNorway.450k[,1:2])
  } 

  # reprocess some TCGA 450k data from IDATs 
  if (require(minfi) && 
    require(TCGAMethylation450k) && 
    require(IlluminaHumanMethylation450kmanifest) && 
    require(IlluminaHumanMethylation450kanno.ilmn12.hg19)) {
    TCGA_path <- system.file("extdata","idat", package="TCGAMethylation450k")
    TCGA_IDATs <- list.files(TCGA_path)
    TCGA_stubs <- unique(sub("_(Red|Grn).idat", "", TCGA_IDATs))
    TCGA_targets <- data.frame(Basename=TCGA_stubs)
    TCGA_rgSet <- read.metharray.exp(base=TCGA_path, targets=TCGA_targets)
    TCGA_grSet <- sesamize(TCGA_rgSet)
  }

}
}
