% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesamizeGEO.R
\name{sesamizeGEO}
\alias{sesamizeGEO}
\title{Sesamize a bunch of IDATs from GEO (or anywhere else with sensible names)}
\usage{
sesamizeGEO(subjects, frags = 1:3, annot = FALSE, HDF5 = FALSE,
  cachePath = NULL, rightType = NULL, force = FALSE,
  check.frags = FALSE, ...)
}
\arguments{
\item{subjects}{The names of each subject, or an rgSet}

\item{frags}{Which elements to extract for the array Basename (DEFAULT: 1:3)}

\item{annot}{Look up the titles and characteristics for GSMs (DEFAULT: FALSE)}

\item{HDF5}{EXPERIMENTAL: make the grSet HDF5-backed? (DEFAULT: FALSE)}

\item{cachePath}{Where to cache the GEOmetadb sqlite file (DEFAULT: tempdir)}

\item{rightType}{"Correct" array type, tries to guess is NULL (DEFAULT: NULL)}

\item{force}{Parse IDAT files with different array size,
see the man page ?read.metharray (DEFAULT: FALSE)}

\item{check.frags}{Check if frags covers all elements up to "idat.gz"
(DEFAULT: FALSE)}

\item{...}{More arguments to pass on to sesamize}
}
\value{
GenomicRatioSet with metadata(grSet)$SNPs filled out
}
\description{
This function processes a directory full of IDATs, somewhat sensibly.
}
\details{
If an RGChannelSet object is provided as the first argument, assume it's
already got a $subject colData column and process as if we'd got to there.

A better-er idea would probably involve running the IDATs individually and
then bolting them all together into an HDF5-backed grSet at the last instant.
}
