% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPcalls.R
\name{SNPcalls}
\alias{SNPcalls}
\title{fit a 1-to-3-component mixture model to flogit(SNPbetas) for each sample}
\usage{
SNPcalls(SNPs)
}
\arguments{
\item{SNPs}{a GenomicRatioSet or a matrix of SNP beta values}
}
\value{
\preformatted{     0/1/2 calls for each SNP in each sample
}
}
\description{
fit a 1-to-3-component mixture model to flogit(SNPbetas) for each sample
}
\details{
\preformatted{    Mclust(miser::flogit(x), G=1:3) is run for each sample. 
}
}
