% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testIDATs.R
\name{testIDATs}
\alias{testIDATs}
\title{Make sure annotations and so forth are in place before doing a big run...}
\usage{
testIDATs(samps, ...)
}
\arguments{
\item{samps}{Two or more rows of the \code{samps} data.frame with \code{Basename}}

\item{...}{Other arguments to be passed to readMethArrays}
}
\value{
Boolean of if sesamize finished successfully
}
\description{
Make sure annotations and so forth are in place before doing a big run...
}
\examples{

}
