% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XYstats.R
\name{XYstats}
\alias{XYstats}
\title{compute assorted sex chromosome related statistics (XNA, YNA, XBeta)}
\usage{
XYstats(grSet)
}
\arguments{
\item{grSet}{a GenomicRatioSet}
}
\value{
\preformatted{     a matrix of XNA, YNA, XBeta for each column
}
}
\description{
compute assorted sex chromosome related statistics (XNA, YNA, XBeta)
}
