% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCharacteristics.R
\name{addCharacteristics}
\alias{addCharacteristics}
\title{get the characteristics of a bunch of GSM entries}
\usage{
addCharacteristics(x, column = "subject", titles = TRUE, cachePath = NULL)
}
\arguments{
\item{x}{an RGChannelSet or GenomicRatioSet with $subject, or GSMs}

\item{column}{name of the column holding the GSMs ("subject")}

\item{titles}{tack on the titles as well? (TRUE)}

\item{cachePath}{where to cache the GEOmetadb sqlite file (tempdir())}
}
\value{
\preformatted{       an rg/grSet, perhaps with more colData, or a data.frame
}
}
\description{
like it says on the tin. this is usually for filling out pData/colData
}
