% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asMethBlocks.R
\name{asMethBlocks}
\alias{asMethBlocks}
\title{Collapse a GenomicRatioSet or similar over methylation blocks (per Kaplan)}
\usage{
asMethBlocks(x, y = NULL, useCpH = FALSE, minprobes = 1, obj = TRUE)
}
\arguments{
\item{x}{usually, a GenomicRatioSet or something like it}

\item{y}{regions to summarize over (methBlocks.EPIC.hg19)}

\item{useCpH}{include CpH probes? (FALSE, because it's a bad idea)}

\item{minprobes}{minimum probes to summarize over (1)}

\item{obj}{keep the object the same (e.g. summarizedExperiment)? (TRUE)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     an object with same colData but with new rowRanges & assays
}\if{html}{\out{</div>}}
}
\description{
Collapse a GenomicRatioSet or similar over methylation blocks (per Kaplan)
}
\details{
By default, methBlocks.EPIC.hg19 is used for summarization, and CpHs are not.
}
