% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byExtremality.R
\name{byExtremality}
\alias{byExtremality}
\title{For array-based DNA methylation, particularly summarized across regions,
we can do better (a lot better) than MAD.  Since we know that}
\usage{
byExtremality(x, k = 500)
}
\arguments{
\item{x}{a matrix of beta values, or a GenomicRatioSet}

\item{k}{how many rows to return (500)}
}
\value{
the most extremal \emph{k} rows of \emph{x} (returns the same class as \emph{x})
}
\description{
max(SD(X_j)) if X_j ~ Beta(a, b) < max(SD(X_j)) if X_j ~ Bernoulli(a/(a+b))
}
\details{
for X having a known mean and SD, hence solvable for a + b by MoM, define

extremality = sd(X_j) / bernoulliSD(mean(X_j))

This function selects the k most extremal rows of x and returns their values.
}
