% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_metrics.R
\name{control_metrics}
\alias{control_metrics}
\title{bead array control processing/QC: get metrics from control probe signals}
\usage{
control_metrics(rgSet, dft = NULL, baseline = 3000, biotin.baseline = 1)
}
\arguments{
\item{rgSet}{an rgSet with control probe intensities and annotation}

\item{dft}{optional single-row data.frame of thresholds (default)}

\item{baseline}{offset for Extension green based background (3000)}

\item{biotin.baseline}{baseline offset for biotin staining probes (1)}
}
\value{
\preformatted{           a matrix (cols = metrics, rows = samples)
}
}
\description{
mostly lifted from Sean Maden's recountmethylation supplement
https://github.com/metamaden/recountmethylationManuscriptSupplement
modified slightly to take advantage of rgSet annotations and to Go Big,
i.e. this works reasonably well on fairly large rgSets (like all of GEO)
}
\examples{
  
if (exists("MPAL_rgSet")) {

  # TARGET MPAL data, from Alexander et al, Nature 2018
  cm <- control_metrics(MPAL_rgSet)

  library(ComplexHeatmap)
  Heatmap(t(cm), name="metric", column_names_side="top",
          column_names_gp=gpar(fontsize=6), row_names_side="left")

  library(circlize)
  flagcols <- colorRamp2(c(0, 1), c("white", "darkred"))
  flagged <- flag_control_failures(cm, platform="epic")
  all_samples_passed <- ifelse(colSums(flagged) == 0, 1, 0)
  all_probes_passed <- ifelse(rowSums(flagged) == 0, 1, 0)
  Heatmap(t(flagged), name="failed", col=flagcols, column_names_side="top",
          column_split=all_probes_passed, column_names_gp=gpar(fontsize=6), 
          row_split=all_samples_passed, row_names_side="left")

} 

}
\seealso{
\preformatted{          flag_control_failures
}
}
