% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpQCfiles.R
\name{dumpQCfiles}
\alias{dumpQCfiles}
\title{dump QC (NAs, flags) and SNP/XY covariates CSVs (optionally, betas as well)}
\usage{
dumpQCfiles(grSet, stub, path = ".", betas = FALSE)
}
\arguments{
\item{grSet}{GenomicRatioSet with betas and QC metrics to dump out}

\item{stub}{mandatory file prefix (_qc, _SNPXY, _betas.csv.gz)}

\item{path}{where to save the dumped CSV file\link{s} (".")}

\item{betas}{dump beta values too? (FALSE)}
}
\value{
a GenomicRatioSet (or an rgSet in case of failure)
}
\description{
SNPs and sex are saved in stub_SNPXY.csv.gz
NA fraction and flagged controls are saved in stub_qc.csv.gz
sesamized betas (including NA-masked values) are saved in stub_betas.csv.gz
(but only if betas == TRUE! this tends to be an enormous file, obviously)
}
