% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elts.R
\name{elts}
\alias{elts}
\title{get elements from a consistently separated character string}
\usage{
elts(x, sep = "_", elt = 1, naify = FALSE)
}
\arguments{
\item{x}{the string(s) to split}

\item{sep}{the separator (default is "_")}

\item{elt}{the element\link{s} to get (default is element 1)}

\item{naify}{return NA_character_ if y is not present? (FALSE)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a string, joined by `sep` if `elt` is a vector
}\if{html}{\out{</div>}}
}
\description{
Another function that I keep rewriting; no more!
}
\examples{

string <- "foo_123_bar.baz.qux"

elts(string)                     # [1] "foo" 
elts(string, elts=2:3)           # [1] "123_bar.baz.qux"
elts(string, sep=".", elts=2:3)  # [1] "baz.qux"

strings <- paste0(letters[1:3], "_string")

elts(strings)              # a_string b_string c_string 
                           #      "a"      "b"      "c" 

strings2 <- c(strings, "nothing") 

elts(strings2)             #  a_string  b_string  c_string   nothing 
                           #       "a"       "b"       "c" "nothing" 
elts(strings2, naify=TRUE) #  a_string b_string c_string  nothing 
                           #       "a"      "b"      "c"       NA 

}
