% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstPass.R
\name{firstPass}
\alias{firstPass}
\title{plot heatmap, remind user to check SNPs}
\usage{
firstPass(x, k = 500, ...)
}
\arguments{
\item{x}{a grSet or matrix of beta values}

\item{k}{how many features to keep (max)}
}
\description{
plot heatmap, remind user to check SNPs
}
