% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSexMix.R
\name{fitSexMix}
\alias{fitSexMix}
\title{simple function for plotting a mixture model for sex from XY stats}
\usage{
fitSexMix(x, ...)
}
\arguments{
\item{x}{a GenomicRatioSet, or an XYstats matrix}

\item{...}{arguments to pass to Mclust}
}
\value{
\preformatted{   an mclust fit object with $sex as fitted sex
}
}
\description{
simple function for plotting a mixture model for sex from XY stats
}
