% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_control_failures.R
\name{flag_control_failures}
\alias{flag_control_failures}
\title{get control quality assessments from an object returned by \code{bactrl}}
\usage{
flag_control_failures(rmat, dft = NULL)
}
\arguments{
\item{rmat}{return matrix from control_metrics(...), N rows by M cols, or
an RGChannelSet (in which case control_metrics(x) will be run)}

\item{dft}{optional threshold data.frame; defaults will be used if null}
}
\value{
\preformatted{ an NxM matrix of threshold assessments, 0 = pass, 1 = fail
}
}
\description{
Apply BeadArray thresholds for quality flagging. Thresholds obtained from
the BeadArray Controls Reporter Software Guide (v00, source 2) and ewastools
resource (v1.5, source 3).
}
