% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBarcodes.R
\name{getBarcodes}
\alias{getBarcodes}
\title{retrieve all the IDAT Basenames in a directory}
\usage{
getBarcodes(dir = ".", z = 1:2, ...)
}
\arguments{
\item{dir}{the path to the directory (default: use ".", the current path)}

\item{z}{which fragments to use as the barcode (default: 1 and 2)}

\item{...}{other arguments to pass to elts()}
}
\value{
\preformatted{ the barcodes
}
}
\description{
Note that getBarcodes(dir, 1) will get the chip address for unmolested IDATs.
}
