% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDMRs.R
\name{getDMRs}
\alias{getDMRs}
\title{use DMRcate to get DMRs (with less hassle)}
\usage{
getDMRs(
  x,
  design,
  dropXY = TRUE,
  impute = FALSE,
  coef = 2,
  fdr = 0.05,
  betacutoff = 0.1,
  DMLs = FALSE,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{x}{a GenomicRatioSet}

\item{design}{a design matrix}

\item{dropXY}{drop sex chromosomes? (TRUE, usually a good idea)}

\item{impute}{impute NAs (FALSE; this will hang on HDF5 if TRUE)}

\item{coef}{which column to fit (default is 2)}

\item{betacutoff}{DMRs must have at least this maxbetaFC (0.1)}

\item{...}{other arguments to pass to DMRcate::cpg.annotate}
}
\value{
\preformatted{           tidied output from DMRcate::dmrcate
}
}
\description{
This is a convenience wrapper that leaves out much of DMRcate's flexibility.
If you need that flexibility, you are well advised to use DMRcate directly.
}
