% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDMRs.R
\name{getDMRs}
\alias{getDMRs}
\title{use DMRcate to get DMRs (with less hassle)}
\usage{
getDMRs(
  x,
  design = NULL,
  dropXY = TRUE,
  impute = TRUE,
  coef = 2,
  fdr = 0.05,
  betacutoff = 0.1,
  DMLs = TRUE,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{x}{a GenomicRatioSet}

\item{design}{a design matrix (NULL)}

\item{dropXY}{drop sex chromosomes? (TRUE, usually a good idea)}

\item{impute}{impute NAs (TRUE, but beware, this WILL hang on HDF5)}

\item{coef}{which column to fit (or "all") (default is 2)}

\item{fdr}{FDR cutoff (0.05)}

\item{betacutoff}{DMRs must have at least this maxbetaFC (0.1)}

\item{DMLs}{keep raw differentially methylated loci to plot (TRUE)}

\item{parallel}{run in parallel? (often unwise; uses mclapply) (FALSE)}

\item{...}{other arguments to pass to dmrcate}
}
\value{
\preformatted{           a dmrcate.output object
}
}
\description{
This is a convenience wrapper that leaves out much of DMRcate's flexibility.
If you need that flexibility, you are well advised to use DMRcate directly.
}
