% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDAT.R
\name{getIDAT}
\alias{getIDAT}
\title{convenience function; checks to see if an IDAT file was downloaded already}
\usage{
getIDAT(IDAT, path = ".")
}
\arguments{
\item{IDAT}{the path to an IDAT (e.g. getGEO(GSE)[\link{1}]$supplemental_file\link{1})}

\item{path}{where to look for it or store it locally (".")}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ the local file path, if present or if downloaded successfully
}\if{html}{\out{</div>}}
}
\description{
convenience function; checks to see if an IDAT file was downloaded already
}
