% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRGChannelSet.R
\name{getRGChannelSet}
\alias{getRGChannelSet}
\title{assemble a suitable RGChannelSet for sesamizing}
\usage{
getRGChannelSet(subjects = NULL, frags = 1:3)
}
\arguments{
\item{subjects}{the names of each subject, or NULL to autodetect (NULL)}

\item{frags}{which elements to extract for the array Basename (1:3)}
}
\value{
\preformatted{       an RGChannelSet with pData $subject and $Basename filled 
}
}
\description{
Note: this should probably go away in favor of signalSet parallel processing
}
