% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSamps.R
\name{getSamps}
\alias{getSamps}
\title{helper function to read in samples and Basename columns}
\usage{
getSamps(subjects = NULL, frags = 1:3)
}
\arguments{
\item{subjects}{the names of each subject, or NULL to autodetect (NULL)}

\item{frags}{which elements to extract for the array Basename (1:3)}
}
\value{
\preformatted{       a data.frame
}
}
\description{
helper function to read in samples and Basename columns
}
