% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMeth.R
\name{mergeMeth}
\alias{mergeMeth}
\title{helper function to merge grSets}
\usage{
mergeMeth(x, y, verbose = TRUE)
}
\arguments{
\item{x}{a GenomicRatioSet}

\item{y}{another GenomicRatioSet}

\item{verbose}{be verbose? (TRUE)}
}
\value{
\preformatted{   a GenomicRatioSet combining the two
}
}
\description{
helper function to merge grSets
}
\examples{
# merged <- Reduce(mergeMeth, list(grs1, grs2, grs3))

}
