% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAfrac.R
\name{NAfrac}
\alias{NAfrac}
\title{compute the fraction of sample-specific (not masked) NA probes}
\usage{
NAfrac(grSet)
}
\arguments{
\item{grSet}{a GenomicRatioSet}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     the fraction NA (of probes that aren't 100\% NA) per column
}\if{html}{\out{</div>}}
}
\description{
compute the fraction of sample-specific (not masked) NA probes
}
