% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFlaggedControls.R
\name{plotFlaggedControls}
\alias{plotFlaggedControls}
\title{simple function for plotting control probe failure flaggings}
\usage{
plotFlaggedControls(grSet, ...)
}
\arguments{
\item{grSet}{GenomicRatioSet where !is.null(metadata(grSet)$control_flagged)}

\item{...}{additional parameters to pass to ComplexHeatmap::Heatmap}
}
\value{
a Heatmap
}
\description{
The data is drawn from metadata(grSet)$control_flagged; if not present,
the function will fail with an error message. See flag_control_failures().
}
