% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSNPs.R
\name{plotSNPs}
\alias{plotSNPs}
\title{another oldie but goodie}
\usage{
plotSNPs(x)
}
\arguments{
\item{x}{a grSet with SNPs in its metadata()}
}
\description{
plot SNPs living in metadata(x)$SNPs, provided they match sampleNames(x)
}
