% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotXYstats.R
\name{plotXYstats}
\alias{plotXYstats}
\title{simple function for plotting XY stats (split by inferred sex if found)}
\usage{
plotXYstats(x, sex = NULL, ...)
}
\arguments{
\item{x}{a GenomicRatioSet, or an XYstats matrix}

\item{sex}{optional sex labels to use for splitting the plot}

\item{...}{additional parameters to pass to ComplexHeatmap::Heatmap}
}
\value{
a Heatmap
}
\description{
simple function for plotting XY stats (split by inferred sex if found)
}
