% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFromIDATs.R
\name{processFromIDATs}
\alias{processFromIDATs}
\title{read in a directory's worth of IDATs, QC them, optionally sesamize them all}
\usage{
processFromIDATs(
  frags = 1:3,
  targets = NULL,
  addgeo = FALSE,
  justRgSet = FALSE,
  forSesame = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{frags}{which elements of the filenames are relevant? (1:3)}

\item{targets}{a targets dataframe (will be passed to minfi for reading)}

\item{addgeo}{optional: try to annotate from GEO? (FALSE)}

\item{justRgSet}{optional: dump the result and don't sesamize? (FALSE)}

\item{forSesame}{optional: avoid loading minfi data structures? (FALSE)}

\item{...}{options to pass to sesamize}
}
\value{
a SummarizedExperiment-derived object of some sort (see Details)
}
\description{
Note: this function simply reads all IDATs in the current directory,
runs QC, stuffs in metadata, and (if !justRgSet) sesamizes it.
}
\details{
minfi is becoming a pain in the ass to deal with due to its dependencies, so
this function can emit various SummarizedExperiment-derived results (whether
an RGChannelSet, a GenomicRatioSet, a SummarizedExperiment, or a
SingleCellExperiment, depending upon its parameters). The only real guarantee
is that, in the absence of an error, a SummarizedExperiment-derived object
will be returned, and it will accommodate the usual generic getWhatever()s.
}
