% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFromIDATs.R
\name{processFromIDATs}
\alias{processFromIDATs}
\title{read in a directory's worth of IDATs, flag controls, and sesamize them all}
\usage{
processFromIDATs(...)
}
\arguments{
\item{...}{options to pass to sesame::sesamize}
}
\value{
a GenomicRatioSet
}
\description{
Note: this function simply reads all IDATs in the current directory,
runs QC, stuffs that into metadata, and sesamizes the result.
}
