% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesamask.R
\name{sesamask}
\alias{sesamask}
\title{add a mask to rowData(grSet) in a platform-appropriate way}
\usage{
sesamask(x)
}
\arguments{
\item{x}{a GenomicRatioSet}
}
\value{
the same GenomicRatioSet but with $mask in rowData(x) filled out
}
\description{
since GenomicRatioSets have their platform annotated, we just use that
}
