% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesamizeGEO.R
\name{sesamizeGEO}
\alias{sesamizeGEO}
\title{sesamize a bunch of IDATs from GEO (or anywhere else with sensible names)}
\usage{
sesamizeGEO(subjects = NULL, elts = 1:3, mask = TRUE,
  titles = FALSE, ...)
}
\arguments{
\item{subjects}{the names of each subject, if known (default: autodetect)}

\item{elts}{which elements to extract for the array Basename (1:3)}

\item{mask}{add rowData(grSet)$mask using sesameData? (TRUE)}

\item{titles}{look up the titles for GSM entries from GEO? (FALSE)}

\item{...}{more arguments to pass on to sesamize}
}
\value{
\preformatted{       a GenomicRatioSet with metadata(grSet)$SNPs filled out
}
}
\description{
This function processes a directory full of IDATs, but sensibly.
}
