% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesamizeGEO.R
\name{sesamizeGEO}
\alias{sesamizeGEO}
\title{sesamize a bunch of IDATs from GEO (or anywhere else with sensible names)}
\usage{
sesamizeGEO(
  subjects = NULL,
  frags = 1:3,
  annot = FALSE,
  HDF5 = FALSE,
  cachePath = NULL,
  ...
)
}
\arguments{
\item{subjects}{the names of each subject, or an rgSet (autodetect)}

\item{frags}{which elements to extract for the array Basename (1:3)}

\item{annot}{look up the titles and characteristics for GSMs? (FALSE)}

\item{HDF5}{EXPERIMENTAL: make the grSet HDF5-backed? (FALSE)}

\item{cachePath}{where to cache the GEOmetadb sqlite file (tempdir())}

\item{...}{more arguments to pass on to sesamize}
}
\value{
\preformatted{       a GenomicRatioSet with metadata(grSet)$SNPs filled out
}
}
\description{
This function processes a directory full of IDATs, somewhat sensibly.
}
\details{
If an RGChannelSet object is provided as the first argument, assume it's
already got a $subject colData column and process as if we'd got to there.

A better-er idea would probably involve running the IDATs individually and
then bolting them all together into an HDF5-backed grSet at the last instant.
}
