% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testIDATs.R
\name{testIDATs}
\alias{testIDATs}
\title{make sure annotations and so forth are in place before doing a big run...}
\usage{
testIDATs(samps, ...)
}
\arguments{
\item{samps}{two or more rows of the \code{samps} data.frame with \code{Basename}}

\item{...}{other arguments, currently unused}
}
\value{
\preformatted{     the status of an attempt to annotate those IDATs 
}
}
\description{
make sure annotations and so forth are in place before doing a big run...
}
