% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vafClust.R
\name{vafClust}
\alias{vafClust}
\title{punishingly simple VAF clustering for clonality estimation}
\usage{
vafClust(alt, ref, depth = NULL, ...)
}
\arguments{
\item{alt}{read depth for the variant allele}

\item{ref}{read depth for the reference allele}

\item{...}{additional arguments to pass on to dbscan}

\item{eps}{epsilon to use for DBSCAN clustering}
}
\value{
\preformatted{   an object of class `dbscan_fast`
}
}
\description{
Suppose we have read coverage over a base with a detected variant.
We can model the reads mapping to variant and reference bases as
}
\details{
p ~ Beta(variant_reads, reference_reads)

Now, if we take logits, with nu = logit(p), then

nu ~ Normal( log(p/(1-p)), log10(total reads) )

As a bonus, we can model \code{nu} as roughly multivariate normal.
Furthermore, we can use the latter term as observation weights.
Given a vector or matrix of VAFs, this function clusters them
using dbscan, with log10(read_depth) as the weight for each VAF.
}
